style: apply laravel pint fixes
