+++ src/Exceptions/InvalidRelationshipException.php
@@ -6,4 +6,3 @@ class InvalidRelationshipException extends MongoAutoSyncException
 {
     //
 }
-
