+++ src/Exceptions/ModelNotFoundException.php
@@ -6,4 +6,3 @@ class ModelNotFoundException extends MongoAutoSyncException
 {
     //
 }
-
